+++ app/profile/allergens.jsx
@@ -13,7 +13,7 @@ import { useRouter } from "expo-router";
 import { useState, useEffect } from "react";
 import { ThemedText } from "@/components/ui/ThemedText";
 import { LinearGradient } from "expo-linear-gradient";
-import { useSettings } from "@/hooks/useSettings";
+import { useSettings } from "@/hooks/useAllergies";
 import Ionicons from "@expo/vector-icons/Ionicons";
 
 export default function AllergensScreen() {
